Move ValidateAIF program into its own class, ValidateAIFCli. Add ValidateAIF.setThreadCount to control multi-threading of validator. Add ValidateCLITest to test validator program. Add picocli library to pom
