+++ pom.xml
@@ -89,6 +89,13 @@
             <version>1.2.0-SNAPSHOT</version>
         </dependency>
 
+        <!-- for validator cli -->
+        <dependency>
+            <groupId>info.picocli</groupId>
+            <artifactId>picocli</artifactId>
+            <version>4.0.0-alpha-3</version>
+        </dependency>
+
         <!-- for testing -->
         <dependency>
             <groupId>org.junit.jupiter</groupId>
@@ -195,7 +202,7 @@
                                 </program>
 
                                    <program>
-                                    <mainClass>com.ncc.aif.ValidateAIF</mainClass>
+                                    <mainClass>com.ncc.aif.ValidateAIFCli</mainClass>
                                     <id>validateAIF</id>
                                 </program>
                             </programs>
