+++ src/test/java/com/ncc/aif/NistTA3ExamplesAndValidationTest.java
@@ -30,7 +30,7 @@ public class NistTA3ExamplesAndValidationTest {
     static void initTest() {
         // prevent too much logging from obscuring the Turtle examples which will be printed
         ((Logger) org.slf4j.LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME)).setLevel(Level.INFO);
-        utils = new NistTA3TestUtils(LDC_NS, ValidateAIF.createForLDCOntology(ValidateAIF.Restriction.NIST_HYPOTHESIS), DUMP_ALWAYS, DUMP_TO_FILE);
+        utils = new NistTA3TestUtils(LDC_NS, ValidateAIF.createForLDCOntology(ValidateAIF.Restriction.NIST_TA3), DUMP_ALWAYS, DUMP_TO_FILE);
     }
 
     private Model model;
